Add import/export/reset settings + refactor SettingsFragment
