+++ app/src/main/java/com/philkes/notallyx/presentation/viewmodel/preference/NotallyXPreferences.kt
@@ -1,6 +1,8 @@
 package com.philkes.notallyx.presentation.viewmodel.preference
 
 import android.app.Application
+import android.content.Context
+import android.net.Uri
 import androidx.core.content.edit
 import androidx.preference.PreferenceManager
 import androidx.security.crypto.EncryptedSharedPreferences
@@ -8,8 +10,13 @@ import androidx.security.crypto.MasterKey
 import com.philkes.notallyx.R
 import com.philkes.notallyx.data.model.Type
 import com.philkes.notallyx.presentation.viewmodel.preference.Constants.PASSWORD_EMPTY
+import com.philkes.notallyx.utils.Operations
+import java.io.IOException
+import java.io.InputStream
+import org.json.JSONArray
+import org.json.JSONObject
 
-class NotallyXPreferences private constructor(app: Application) {
+class NotallyXPreferences private constructor(private val app: Application) {
 
     private val preferences = PreferenceManager.getDefaultSharedPreferences(app)
 
@@ -148,6 +155,89 @@ class NotallyXPreferences private constructor(app: Application) {
 
     fun showDateCreated(): Boolean {
         return dateFormat.value != DateFormat.NONE
+    }
+
+    fun export(context: Context, uri: Uri): Boolean {
+        val jsonObject = JSONObject()
+        try {
+            for ((key, value) in preferences.all) {
+                if (key in listOf(biometricLock.key, iv.key, databaseEncryptionKey.key)) {
+                    continue
+                }
+                when (value) {
+                    is Collection<*> -> jsonObject.put(key, JSONArray(value))
+                    is Enum<*> -> jsonObject.put(key, value.name.toCamelCase())
+                    else -> jsonObject.put(key, value)
+                }
+            }
+            context.contentResolver.openOutputStream(uri)?.use {
+                it.write(jsonObject.toString(4).toByteArray())
+            } ?: return false
+            return true
+        } catch (e: IOException) {
+            Operations.log(app, e)
+            return false
+        }
+    }
+
+    fun import(context: Context, uri: Uri): Boolean {
+        try {
+            val inputStream: InputStream? = context.contentResolver.openInputStream(uri)
+            val jsonString = inputStream?.bufferedReader()?.use { it.readText() } ?: return false
+
+            val jsonObject = JSONObject(jsonString)
+
+            val editor = preferences.edit()
+            editor.clear()
+
+            jsonObject.keys().forEach { key ->
+                when (val value = jsonObject.get(key)) {
+                    is Int -> editor.putInt(key, value)
+                    is Boolean -> editor.putBoolean(key, value)
+                    is Double -> editor.putFloat(key, value.toFloat())
+                    is Long -> editor.putLong(key, value)
+                    is JSONArray -> {
+                        val set = (0 until value.length()).map { value.getString(it) }.toSet()
+                        editor.putStringSet(key, set)
+                    }
+                    else -> editor.putString(key, value.toString())
+                }
+            }
+            editor.apply()
+            reload()
+            return true
+        } catch (e: Exception) {
+            Operations.log(app, e)
+            return false
+        }
+    }
+
+    fun reset() {
+        preferences.edit().clear().apply()
+        encryptedPreferences.edit().clear().apply()
+        reload()
+    }
+
+    private fun reload() {
+        setOf(
+                theme,
+                textSize,
+                notesView,
+                listItemSorting,
+                labelsHiddenInNavigation,
+                autoBackup,
+                dataOnExternalStorage,
+                dateFormat,
+                maxItems,
+                maxTitle,
+                maxLabels,
+                maxLines,
+                notesSorting,
+                backupPassword,
+                listItemSorting,
+                biometricLock,
+            )
+            .forEach { it.refresh() }
     }
 
     companion object {
